refactor(graph): extract edge semantic pictogram from link map IIFE
PR review: encapsulate EdgeSemanticPictogram logic in GraphLinkEdgeSemanticPictogram
for D3ForceGraph readability. Remove unused isExistingContextLink.
